새로 생성한 RIB Package 세팅
