+++ Profile/Package.swift
@@ -24,6 +24,12 @@ let package = Package(
         .library(
             name: "ProfileDesignSystem",
             targets: ["ProfileDesignSystem"]),
+        .library(
+            name: "ProfileRestore",
+            targets: ["ProfileRestore"]),
+        .library(
+            name: "ProfileBackup",
+            targets: ["ProfileBackup"]),
     ],
     dependencies: [
         .package(url: "https://github.com/uber/RIBs", branch: "main"),
@@ -73,8 +79,7 @@ let package = Package(
                 .product(name: "RxCocoa", package: "RxSwift"),
                 .product(name: "DesignSystem", package: "Platform"),
                 .product(name: "MenualEntity", package: "Platform"),
-                .product(name: "MenualRepository", package: "Platform"),
-                .product(name: "ZipArchive", package: "ZipArchive")
+                .product(name: "MenualRepository", package: "Platform")
             ]
         ),
         .target(
@@ -91,6 +96,23 @@ let package = Package(
         ),
         .target(
             name: "ProfileDesignSystem",
+            dependencies: [
+                .product(name: "RIBs", package: "RIBs"),
+                .product(name: "DesignSystem", package: "Platform"),
+                .product(name: "MenualEntity", package: "Platform"),
+            ]
+        ),
+        .target(
+            name: "ProfileBackup",
+            dependencies: [
+                .product(name: "RIBs", package: "RIBs"),
+                .product(name: "DesignSystem", package: "Platform"),
+                .product(name: "MenualEntity", package: "Platform"),
+                .product(name: "ZipArchive", package: "ZipArchive"),
+            ]
+        ),
+        .target(
+            name: "ProfileRestore",
             dependencies: [
                 .product(name: "RIBs", package: "RIBs"),
                 .product(name: "DesignSystem", package: "Platform"),
