+++ Profile/Sources/ProfileBackup/ProfileBackupInteractor.swift
@@ -7,6 +7,7 @@
 
 import RIBs
 import RxSwift
+import ZipArchive
 
 protocol ProfileBackupRouting: ViewableRouting {
     // TODO: Declare methods the interactor can invoke to manage sub-tree via the router.
@@ -15,6 +16,7 @@ protocol ProfileBackupRouting: ViewableRouting {
 protocol ProfileBackupPresentable: Presentable {
     var listener: ProfileBackupPresentableListener? { get set }
     // TODO: Declare methods the interactor can invoke the presenter to present data.
+    func showShareSheet(path: String)
 }
 
 protocol ProfileBackupListener: AnyObject {
@@ -41,5 +43,51 @@ final class ProfileBackupInteractor: PresentableInteractor<ProfileBackupPresenta
     override func willResignActive() {
         super.willResignActive()
         // TODO: Pause any business logic.
+    }
+    
+    func tempZipPath() -> String {
+        var path = NSSearchPathForDirectoriesInDomains(.cachesDirectory, .userDomainMask, true)[0]
+        path += "/\(UUID().uuidString).zip"
+        return path
+    }
+    
+    func saveiCloud() {
+        let documentDirectory = FileManager.SearchPathDirectory.documentDirectory
+        let userDomainMask = FileManager.SearchPathDomainMask.userDomainMask
+        let path = NSSearchPathForDirectoriesInDomains(documentDirectory, userDomainMask, true)
+
+        var filePaths: [String] = []
+        var fileElements: [String] = []
+        var folderPaths: [String] = []
+        let enumerator = FileManager.default.enumerator(atPath: path.first!)
+        // Documents폴더를 돌면서 파일의 절대값과, 업로드하고자 하는 상대값을 두 Array에 담아서 저장
+        // 폴더는 미리 생성하기 위해서 저장
+        while let element = enumerator?.nextObject() as? String {
+            if let fType = enumerator?.fileAttributes?[FileAttributeKey.type] as? FileAttributeType {
+                switch fType {
+                case .typeRegular:
+                    filePaths.append(path.first! + "/" + element)
+                    fileElements.append(element)
+                case .typeDirectory:
+                    folderPaths.append(element)
+                default:
+                    break
+                }
+            }
+        }
+
+        let savePath: String = tempZipPath()
+        let zip = SSZipArchive.init(path: savePath)
+        zip.open()
+        for path in folderPaths {
+            zip.writeFolder(atPath: savePath, withFolderName: path, withPassword: nil)
+        }
+
+        for (index, file) in filePaths.enumerated() {
+            zip.writeFile(atPath: file, withFileName: fileElements[index], withPassword: nil)
+        }
+        zip.close()
+
+        presenter.showShareSheet(path: savePath)
     }
 }
