+++ Profile/Sources/ProfileBackup/ProfileBackupViewController.swift
@@ -19,3 +19,23 @@ final class ProfileBackupViewController: UIViewController, ProfileBackupPresenta
 
     weak var listener: ProfileBackupPresentableListener?
 }
+
+// MARK: - 파일 공유
+extension ProfileBackupViewController: UIDocumentPickerDelegate {
+    func showShareSheet(path: String) {
+        print("ProfileHome :: path! = \(path)")
+        let fileURL = NSURL(fileURLWithPath: path)
+
+        // Create the Array which includes the files you want to share
+        var filesToShare = [Any]()
+
+        // Add the path of the file to the Array
+        filesToShare.append(fileURL)
+
+        // Make the activityViewContoller which shows the share-view
+        let activityViewController = UIActivityViewController(activityItems: filesToShare, applicationActivities: nil)
+
+        // Show the share-view
+        self.present(activityViewController, animated: true, completion: nil)
+    }
+}
