+++ Profile/Sources/ProfileHome/ProfileHomeInteractor.swift
@@ -9,7 +9,6 @@ import RIBs
 import RxSwift
 import RxRelay
 import RealmSwift
-import ZipArchive
 import DesignSystem
 import MenualRepository
 import MenualEntity
@@ -28,7 +27,6 @@ public protocol ProfileHomeRouting: ViewableRouting {
 protocol ProfileHomePresentable: Presentable {
     var listener: ProfileHomePresentableListener? { get set }
     // TODO: Declare methods the interactor can invoke the presenter to present data.
-    func showShareSheet(path: String)
 }
 
 public protocol ProfileHomeListener: AnyObject {
@@ -215,56 +213,8 @@ final class ProfileHomeInteractor: PresentableInteractor<ProfileHomePresentable>
     
     func goDiaryHome() { }
     
-    // MARK: - iCloud 동기화하기
-    func saveiCloud() {
-        let documentDirectory = FileManager.SearchPathDirectory.documentDirectory
-        let userDomainMask = FileManager.SearchPathDomainMask.userDomainMask
-        let path = NSSearchPathForDirectoriesInDomains(documentDirectory, userDomainMask, true)
-
-        var filePaths: [String] = []
-        var fileElements: [String] = []
-        var folderPaths: [String] = []
-        let enumerator = FileManager.default.enumerator(atPath: path.first!)
-        // Documents폴더를 돌면서 파일의 절대값과, 업로드하고자 하는 상대값을 두 Array에 담아서 저장
-        // 폴더는 미리 생성하기 위해서 저장
-        while let element = enumerator?.nextObject() as? String {
-            if let fType = enumerator?.fileAttributes?[FileAttributeKey.type] as? FileAttributeType {
-                switch fType {
-                case .typeRegular:
-                    filePaths.append(path.first! + "/" + element)
-                    fileElements.append(element)
-                case .typeDirectory:
-                    folderPaths.append(element)
-                default:
-                    break
-                }
-            }
-        }
-
-        let savePath: String = tempZipPath()
-        let zip = SSZipArchive.init(path: savePath)
-        zip.open()
-        for path in folderPaths {
-            zip.writeFolder(atPath: savePath, withFolderName: path, withPassword: nil)
-        }
-
-        for (index, file) in filePaths.enumerated() {
-            zip.writeFile(atPath: file, withFileName: fileElements[index], withPassword: nil)
-        }
-        zip.close()
-
-        presenter.showShareSheet(path: savePath)
-    }
-    
-    func tempZipPath() -> String {
-        var path = NSSearchPathForDirectoriesInDomains(.cachesDirectory, .userDomainMask, true)[0]
-        path += "/\(UUID().uuidString).zip"
-        return path
-    }
-    
     // MARK: - Backup 관련 기능
     func backup() {
         print("ProfileHome :: Backup!")
-        self.saveiCloud()
     }
 }
